+++ magpie/cli/register_defaults.py
@@ -20,7 +20,7 @@ from magpie.utils import get_logger, print_log, raise_log
 
 if TYPE_CHECKING:
     # pylint: disable=W0611,unused-import
-    from typing import Any, AnyStr, Optional, Sequence
+    from typing import Any, Optional, Sequence
 
     from magpie.typedefs import AnySettingsContainer, Str
 
@@ -146,7 +146,7 @@ def init_users_group(db_session, settings=None):
 
 
 def register_defaults(db_session=None, settings=None, ini_file_path=None):
-    # type: (Optional[Session], Optional[AnySettingsContainer], Optional[AnyStr]) -> None
+    # type: (Optional[Session], Optional[AnySettingsContainer], Optional[Str]) -> None
     """
     Registers into database every undefined default users and groups matching following variables:
 
@@ -178,7 +178,7 @@ def make_parser():
 
 
 def main(args=None, parser=None, namespace=None):
-    # type: (Optional[Sequence[AnyStr]], Optional[argparse.ArgumentParser], Optional[argparse.Namespace]) -> Any
+    # type: (Optional[Sequence[Str]], Optional[argparse.ArgumentParser], Optional[argparse.Namespace]) -> Any
     if not parser:
         parser = make_parser()
     args = parser.parse_args(args=args, namespace=namespace)